requirements.txt
pyTelegramBotAPI==4.0.0
requests==2.25.1